tests/requirements.txt
aiohttp==3.8.6
pydantic-settings==2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
redis==5.0.4
pytest-html>=3.2
pytest-metadata>=2.0.0
python-dotenv==1.1.0